fixing names
